Add files via upload
Building a House in Rainfall.
USED ONLY GL_POINTS, GL_LINES, and GL_TRIANGLES for designing the house. The "L" letter will gradually bend the rain to the left, the "R" letter will
gradually bend the rain to the right and the "S" letter will straighten the rain flow. Also, the "D" letter will whiten the background (day) and the "N" letter will darken the background (night).